// src/app/model/asset-information.ts
export class AssetInformation {
    id: number = undefined;
    name: string = undefined;
    key: string = undefined
    description: string = undefined;
    img: string = undefined;
    number: string = undefined;
    requestedFor: string = undefined;
    opened: string = undefined;
    status: string = undefined;
}